2025-2/IA/SMA/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0
aiofiles>=23.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
pydantic>=2.8.0
pydantic-settings>=2.1.0

# Nota: Versões específicas podem causar problemas de compilação com Python 3.13
# Use versões mais recentes que têm wheels pré-compilados disponíveis

